requirements.txt
streamlit
faster-whisper
ffmpeg-python
google-genai
google-cloud-storage
pymupdf-layout
pymupdf4llm
requests
python-dotenv

openpyxl